+++ .adonisjs/server/events.ts
@@ -3,12 +3,14 @@
  * DO NOT EDIT manually
  */
 
+import AiTokenThresholdCrossed from '#events/ai_token_threshold_crossed'
 import OrganizationMemberJoined from '#events/organization_member_joined'
 import OrganizationPlanDowngraded from '#events/organization_plan_downgraded'
 import SimulatorLeadCreated from '#events/simulator_lead_created'
 import StorageThresholdCrossed from '#events/storage_threshold_crossed'
 
 export const events = {
+  AiTokenThresholdCrossed: AiTokenThresholdCrossed,
   OrganizationMemberJoined: OrganizationMemberJoined,
   OrganizationPlanDowngraded: OrganizationPlanDowngraded,
   SimulatorLeadCreated: SimulatorLeadCreated,